+++ packages/web/src/state/reducer.ts
@@ -26,6 +26,8 @@ import { settingsReducer } from './settings/settings.reducer'
 import { uiReducer } from './ui/ui.reducer'
 import { UiState } from './ui/ui.state'
 
+import { errorReducer, ErrorState } from './error/error.reducer'
+
 import { auspiceGeneralReducer, auspiceQueryReducer } from './auspice/auspice.reducer'
 
 export interface State extends AuspiceState {
@@ -33,6 +35,7 @@ export interface State extends AuspiceState {
   settings: SettingsState & PersistedState
   router: RouterState
   ui: UiState
+  error: ErrorState
 }
 
 const SETTINGS_VERSION = 1
@@ -47,6 +50,7 @@ const rootReducer = () =>
     settings: settingsReducerPersisted,
     router: routerReducer,
     ui: uiReducer,
+    error: errorReducer,
 
     // BEGIN reducers from auspice
     metadata,